requirements.txt
torch
torchvision
matplotlib
numpy
tqdm
scikit-learn
seaborn
Pillow